Refactor duplication to use generateVariantFunction
